some bugfixes
