+++ dace/transformation/dataflow/__init__.py
@@ -26,7 +26,6 @@ from .stream_transient import StreamTransient, AccumulateTransient
 from .local_storage import InLocalStorage, OutLocalStorage
 from .double_buffering import DoubleBuffering
 from .streaming_memory import StreamingMemory, StreamingComposition
-from .memory_buffering import MemoryBuffering
 
 # Complexity reduction
 from .dedup_access import DeduplicateAccess
